fix: 🔍 Show the dock icon when the browser window is open
This change toggles the application policy from 'accessory' to 'regular' whenever one or more browser windows are open. This means that whenever a window is visible, the app should appear in the application switcher (Cmd+Tab).

There are also a couple of drive-by fixes, cleaning up the build script a little, and extracting the update view and model into their own files.
