+++ Reconnect/Views/CheckForUpdatesView.swift
@@ -18,20 +18,24 @@
 
 import SwiftUI
 
-struct ContentView: View {
+import Sparkle
 
-    var applicationModel: ApplicationModel
-    let fileServer = FileServer(host: "127.0.0.1", port: 7501)
+// This is the view for the Check for Updates menu item
+// Note this intermediate view is necessary for the disabled state on the menu item to work properly before Monterey.
+// See https://stackoverflow.com/questions/68553092/menu-not-updating-swiftui-bug for more info
+struct CheckForUpdatesView: View {
+    @ObservedObject private var checkForUpdatesViewModel: CheckForUpdatesViewModel
+    private let updater: SPUUpdater
 
-    var body: some View {
-        VStack {
-            if applicationModel.isConnected {
-                BrowserView(fileServer: fileServer)
-            } else {
-                ContentUnavailableView("Not Connected", systemImage: "star")
-            }
-        }
-        .showsDockIcon()
+    init(updater: SPUUpdater) {
+        self.updater = updater
+
+        // Create our view model for our CheckForUpdatesView
+        self.checkForUpdatesViewModel = CheckForUpdatesViewModel(updater: updater)
     }
 
+    var body: some View {
+        Button("Check for Updates…", action: updater.checkForUpdates)
+            .disabled(!checkForUpdatesViewModel.canCheckForUpdates)
+    }
 }
